# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EncoreHub Gateway API",
    "version": "0.1.0",
    "description": "Browser-facing HTTP and SSE contract used by the EncoreHub frontend. Engine-only routes and third-party provider APIs are outside this contract. Every response includes X-Request-ID."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Default headless development address; desktop ports are negotiated at runtime."
    }
  ],
  "security": [
    {},
    {
      "GatewayBearer": []
    }
  ],
...
  "paths": {
...
    "/api/v1/conversations": {
      "get": {
        "tags": ["Conversations"],
        "operationId": "listConversations",
        "responses": {
          "200": { "$ref": "#/components/responses/ConversationList" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Conversations"],
        "operationId": "createConversation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateConversationRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Conversation" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/conversations/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" }
      ],
      "get": {
        "tags": ["Conversations"],
        "operationId": "getConversation",
        "responses": {
          "200": { "$ref": "#/components/responses/ConversationDetail" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["Conversations"],
        "operationId": "updateConversation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateConversationRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Conversation" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Conversations"],
        "operationId": "deleteConversation",
        "responses": {
          "204": { "$ref": "#/components/responses/NoContent" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/v1/conversations/{id}/attachments": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" }
      ],
      "get": {
        "tags": ["Attachments"],
        "operationId": "listConversationAttachments",
        "responses": {
          "200": {
            "description": "Attachment metadata in upload order",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Attachment" } }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Attachments"],
        "operationId": "uploadConversationAttachment",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": { "file": { "type": "string", "format": "binary" } }
              }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/Attachment" },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "413": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/conversations/{id}/attachments/{attachment_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" },
        { "$ref": "#/components/parameters/AttachmentId" }
      ],
      "get": {
        "tags": ["Attachments"],
        "operationId": "getConversationAttachment",
        "responses": {
          "200": { "$ref": "#/components/responses/Attachment" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Attachments"],
        "operationId": "deleteConversationAttachment",
        "responses": {
          "204": { "$ref": "#/components/responses/NoContent" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/conversations/{id}/attachments/{attachment_id}/content": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" },
        { "$ref": "#/components/parameters/AttachmentId" }
      ],
      "get": {
        "tags": ["Attachments"],
        "operationId": "getConversationAttachmentContent",
        "responses": {
          "200": {
            "description": "Original attachment bytes with their recorded content type",
            "content": { "application/octet-stream": { "schema": { "type": "string", "format": "binary" } } }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/conversations/{id}/attachments/{attachment_id}/ocr": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" },
        { "$ref": "#/components/parameters/AttachmentId" }
      ],
      "post": {
        "tags": ["Attachments"],
        "operationId": "ocrConversationAttachment",
        "responses": {
          "200": { "$ref": "#/components/responses/Attachment" },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/v1/conversations/{id}/title": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" }
      ],
      "patch": {
        "tags": ["Conversations"],
        "operationId": "updateConversationTitle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RenameConversationRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Conversation" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/conversations/{id}/messages/{message_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" },
        { "name": "message_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
      ],
      "delete": {
        "tags": ["Conversations"],
        "operationId": "deleteConversationMessage",
        "responses": {
          "204": { "$ref": "#/components/responses/NoContent" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/conversations/{id}/character-upgrade": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" }
      ],
      "get": {
        "tags": ["Characters"],
        "operationId": "previewConversationCharacterUpgrade",
        "responses": {
          "200": { "$ref": "#/components/responses/CharacterUpgradePreview" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Characters"],
        "operationId": "upgradeConversationCharacter",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpgradeCharacterRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Conversation" },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/characters": {
      "get": {
        "tags": ["Characters"],
        "operationId": "listCharacters",
        "responses": {
          "200": { "$ref": "#/components/responses/CharacterList" }
        }
      },
      "post": {
        "tags": ["Characters"],
        "operationId": "createCharacter",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCharacterRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/Character" },
          "400": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/characters/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ResourceId" }
      ],
      "get": {
        "tags": ["Characters"],
        "operationId": "getCharacter",
        "responses": {
          "200": { "$ref": "#/components/responses/Character" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["Characters"],
        "operationId": "updateCharacter",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCharacterRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Character" },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Characters"],
        "operationId": "deleteCharacter",
        "description": "Soft-deletes a non-default character. Existing conversations retain their immutable character snapshots.",
        "responses": {
          "204": { "$ref": "#/components/responses/NoContent" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/v1/characters/{id}/history": {
      "parameters": [{ "$ref": "#/components/parameters/ResourceId" }],
      "get": {
        "tags": ["Characters"],
        "operationId": "getCharacterHistory",
        "responses": {
          "200": {
            "description": "Character branches and immutable version nodes",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CharacterHistory" } } }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/characters/{id}/versions": {
      "parameters": [{ "$ref": "#/components/parameters/ResourceId" }],
      "post": {
        "tags": ["Characters"],
        "operationId": "commitCharacterVersion",
        "description": "Creates an immutable version from the current working copy and advances the active branch.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommitCharacterVersionRequest" } } } },
        "responses": {
          "200": { "$ref": "#/components/responses/Character" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/characters/{id}/versions/{version}/restore": {
      "parameters": [
        { "$ref": "#/components/parameters/ResourceId" },
        { "name": "version", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }
      ],
      "post": {
        "tags": ["Characters"],
        "operationId": "restoreCharacterVersion",
        "description": "Loads an immutable version into the working copy without creating or deleting a version node.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CharacterRevisionRequest" } } } },
        "responses": {
          "200": { "$ref": "#/components/responses/Character" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/characters/{id}/branches": {
      "parameters": [{ "$ref": "#/components/parameters/ResourceId" }],
      "post": {
        "tags": ["Characters"],
        "operationId": "createCharacterBranch",
        "description": "Creates and checks out a branch from an immutable version node.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateCharacterBranchRequest" } } } },
        "responses": {
          "201": { "$ref": "#/components/responses/Character" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/v1/usage": {
      "get": {
        "tags": ["Usage"],
        "operationId": "getUsageReport",
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["15m", "30m", "1h", "3h", "day", "week", "3w", "month", "quarter", "year", "custom", "all"],
              "default": "day"
            }
          },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "provider", "in": "query", "schema": { "type": "string" } },
          { "name": "model", "in": "query", "schema": { "type": "string" } },
          { "name": "currency", "in": "query", "description": "Display currency; defaults to USD. Built-in rates support USD, CNY, and EUR.", "schema": { "type": "string", "default": "USD" } },
          { "name": "timezone_offset_minutes", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Engine-aggregated usage report",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsageReport" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/v1/config/{key}": {
      "parameters": [
        { "$ref": "#/components/parameters/ConfigKey" }
      ],
      "get": {
        "tags": ["Settings"],
        "operationId": "getConfigValue",
        "description": "Returns an Engine-backed JSON configuration value, or null when the key is unset.",
        "responses": {
          "200": {
            "description": "Opaque JSON configuration value",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Settings"],
        "operationId": "setConfigValue",
        "description": "Stores an opaque JSON configuration value. Secrets must use the Secrets API instead.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "204": { "$ref": "#/components/responses/NoContent" }
        }
      }
    },
...
    "/api/v1/skills/match": {
      "get": {
        "tags": ["Skills"],
        "operationId": "matchSkills",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching skills",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SkillList" }
              }
            }
          }
        }
      }
    },
    "/api/v1/skills/{id}/toggle": {
      "parameters": [
        { "$ref": "#/components/parameters/ResourceId" }
      ],
      "post": {
        "tags": ["Skills"],
        "operationId": "toggleSkill",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["enabled"],
                "properties": { "enabled": { "type": "boolean" } }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/NoContent" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/memories": {
      "get": {
        "tags": ["Memories"],
        "operationId": "listMemories",
        "parameters": [
          { "name": "scope", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Persisted memories",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MemoryList" }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/search": {
      "get": {
        "tags": ["Memories"],
        "operationId": "searchMemories",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "scope", "in": "query", "schema": { "type": "string" } },
          { "name": "top_k", "in": "query", "schema": { "type": "integer", "minimum": 1 } }
        ],
        "responses": {
          "200": {
            "description": "Ranked memories",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MemorySearchResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ResourceId" }
      ],
      "delete": {
        "tags": ["Memories"],
        "operationId": "deleteMemory",
        "responses": {
          "204": { "$ref": "#/components/responses/NoContent" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/knowledge": {
      "get": {
        "tags": ["Knowledge"],
        "operationId": "listKnowledgeDocuments",
        "responses": {
          "200": {
            "description": "Knowledge documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/KnowledgeDocument" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Knowledge"],
        "operationId": "ingestKnowledgeDocument",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/KnowledgeIngestRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created knowledge document",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/KnowledgeDocument" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/knowledge/search": {
      "get": {
        "tags": ["Knowledge"],
        "operationId": "searchKnowledge",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "top_k", "in": "query", "schema": { "type": "integer", "minimum": 1 } }
        ],
        "responses": {
          "200": {
            "description": "Ranked knowledge chunks",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/KnowledgeSearchResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ResourceId" }
      ],
      "delete": {
        "tags": ["Knowledge"],
        "operationId": "deleteKnowledgeDocument",
        "responses": {
          "204": { "$ref": "#/components/responses/NoContent" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/v1/secrets": {
      "get": {
        "tags": ["Secrets"],
        "operationId": "listSecretProviders",
        "responses": {
          "200": {
            "description": "Provider ids with stored keys",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SecretsList" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Secrets"],
        "operationId": "putProviderSecret",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PutSecretRequest" }
            }
          }
        },
        "responses": {
          "204": { "$ref": "#/components/responses/NoContent" },
          "423": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/secrets/{provider_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/SecretProviderId" }
      ],
      "get": {
        "tags": ["Secrets"],
        "operationId": "getProviderSecret",
        "responses": {
          "200": {
            "description": "Decrypted provider key",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SecretValue" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" },
          "423": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Secrets"],
        "operationId": "deleteProviderSecret",
        "responses": {
          "204": { "$ref": "#/components/responses/NoContent" },
          "423": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/secrets/enable": {
      "post": {
        "tags": ["Secrets"],
        "operationId": "enableSecretsEncryption",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EnableSecretsRequest" }
            }
          }
        },
        "responses": {
          "204": { "$ref": "#/components/responses/NoContent" },
          "400": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/secrets/disable": {
      "post": {
        "tags": ["Secrets"],
        "operationId": "disableSecretsEncryption",
        "requestBody": { "$ref": "#/components/requestBodies/Password" },
        "responses": {
          "204": { "$ref": "#/components/responses/NoContent" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "GatewayBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Required only when ENCOREHUB_AUTH_TOKEN is configured. Never reuse ENCOREHUB_ENGINE_AUTH_TOKEN."
      }
    },
...
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "NoContent": { "description": "Operation completed" },
...
      "ConversationList": {
        "description": "Conversation summaries",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ConversationList" }
          }
        }
      },
...
    },
    "schemas": {
...
      "CreateConversationRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "provider": { "type": "string" },
          "model": { "type": "string" },
          "character_id": {
            "type": "string",
            "description": "Defaults to the stable default character. Provider/model inherit from the selected character when omitted together."
          }
        }
      },
...
    }
  }
}